requirements.txt
numpy
opencv-python-headless
deskew
black
shapely
msgspec
pillow

pika
boto3
aio-pika
aioboto3
SQLAlchemy
alembic
python-dotenv
motor
pydantic-settings

uuid7
pytz

rapidocr